move telemetry related code to telemetry package (#38690)
* move telemetry related code to telemetry package


* add files
