+++ pilot/pkg/networking/util/util.go
@@ -39,10 +39,8 @@ import (
 	"istio.io/istio/pilot/pkg/features"
 	"istio.io/istio/pilot/pkg/model"
 	istionetworking "istio.io/istio/pilot/pkg/networking"
-	"istio.io/istio/pilot/pkg/serviceregistry/provider"
 	"istio.io/istio/pkg/cluster"
 	"istio.io/istio/pkg/config"
-	"istio.io/istio/pkg/config/host"
 	"istio.io/istio/pkg/config/labels"
 	"istio.io/istio/pkg/network"
 	"istio.io/istio/pkg/proto/merge"
@@ -90,13 +88,6 @@ const (
 	// EnvoyQUICSocketName matched with hardcoded built-in Envoy transport name which determines endpoint
 	// level QUIC transport socket configuration
 	EnvoyQUICSocketName = wellknown.TransportSocketQuic
-
-	// StatName patterns
-	serviceStatPattern         = "%SERVICE%"
-	serviceFQDNStatPattern     = "%SERVICE_FQDN%"
-	servicePortStatPattern     = "%SERVICE_PORT%"
-	servicePortNameStatPattern = "%SERVICE_PORT_NAME%"
-	subsetNameStatPattern      = "%SUBSET_NAME%"
 
 	// Well-known header names
 	AltSvcHeader = "alt-svc"
@@ -553,35 +544,6 @@ func IsAllowAnyOutbound(node *model.Proxy) bool {
 		node.SidecarScope.OutboundTrafficPolicy.Mode == networking.OutboundTrafficPolicy_ALLOW_ANY
 }
 
-// BuildStatPrefix builds a stat prefix based on the stat pattern.
-func BuildStatPrefix(statPattern string, host string, subset string, port *model.Port, attributes *model.ServiceAttributes) string {
-	prefix := strings.ReplaceAll(statPattern, serviceStatPattern, shortHostName(host, attributes))
-	prefix = strings.ReplaceAll(prefix, serviceFQDNStatPattern, host)
-	prefix = strings.ReplaceAll(prefix, subsetNameStatPattern, subset)
-	prefix = strings.ReplaceAll(prefix, servicePortStatPattern, strconv.Itoa(port.Port))
-	prefix = strings.ReplaceAll(prefix, servicePortNameStatPattern, port.Name)
-	return prefix
-}
-
-// BuildInboundStatPrefix builds a stat prefix based on the stat pattern.
-func BuildInboundStatPrefix(statPattern string, tm TelemetryMetadata, subset string, port uint32, portName string) string {
-	prefix := strings.ReplaceAll(statPattern, serviceStatPattern, tm.ShortHostname())
-	prefix = strings.ReplaceAll(prefix, serviceFQDNStatPattern, tm.InstanceHostname.String())
-	prefix = strings.ReplaceAll(prefix, subsetNameStatPattern, subset)
-	prefix = strings.ReplaceAll(prefix, servicePortStatPattern, strconv.Itoa(int(port)))
-	prefix = strings.ReplaceAll(prefix, servicePortNameStatPattern, portName)
-	return prefix
-}
-
-// shortHostName constructs the name from kubernetes hosts based on attributes (name and namespace).
-// For other hosts like VMs, this method does not do any thing - just returns the passed in host as is.
-func shortHostName(host string, attributes *model.ServiceAttributes) string {
-	if attributes.ServiceRegistry == provider.Kubernetes {
-		return attributes.Name + "." + attributes.Namespace
-	}
-	return host
-}
-
 func StringToExactMatch(in []string) []*matcher.StringMatcher {
 	if len(in) == 0 {
 		return nil
@@ -719,7 +681,7 @@ func ByteCount(b int) string {
 		float64(b)/float64(div), "kMGTPE"[exp])
 }
 
-// IPv6 addresses are enclosed in square brackets followed by port number in Host header/URIs
+// IPv6Compliant encloses ipv6 addresses in square brackets followed by port number in Host header/URIs
 func IPv6Compliant(host string) string {
 	if strings.Contains(host, ":") {
 		return "[" + host + "]"
@@ -730,33 +692,4 @@ func IPv6Compliant(host string) string {
 // DomainName builds the domain name for a given host and port
 func DomainName(host string, port int) string {
 	return net.JoinHostPort(host, strconv.Itoa(port))
-}
-
-// TraceOperation builds the string format: "%s:%d/*" for a given host and port
-func TraceOperation(host string, port int) string {
-	// Format : "%s:%d/*"
-	return DomainName(host, port) + "/*"
-}
-
-// TelemetryMetadata defines additional filter chain metadata for telemetry use
-type TelemetryMetadata struct {
-	// InstanceHostname defines the hostname of the service this filter chain is built for.
-	// Note: This is best effort; this may be empty if generated by Sidecar config, and there may be multiple
-	// Services that make up the filter chain.
-	InstanceHostname host.Name
-	// KubernetesServiceNamespace is the namespace the service is defined in, if it is for a Kubernetes Service.
-	// Note: This is best effort; this may be empty if generated by Sidecar config, and there may be multiple
-	// Services that make up the filter chain.
-	KubernetesServiceNamespace string
-	// KubernetesServiceName is the name of service, if it is for a Kubernetes Service.
-	// Note: This is best effort; this may be empty if generated by Sidecar config, and there may be multiple
-	// Services that make up the filter chain.
-	KubernetesServiceName string
-}
-
-func (tm TelemetryMetadata) ShortHostname() string {
-	if tm.KubernetesServiceName != "" {
-		return tm.KubernetesServiceName + "." + tm.KubernetesServiceNamespace
-	}
-	return tm.InstanceHostname.String()
 }
